requirements.txt
streamlit
opencv-python
numpy
pandas
matplotlib
scikit-learn
tensorflow==2.12.0
Pillow
